rename: add -> newproject
- feature.add -> feature.newproject
- AddScreen -> NewProjectScreen
