+++ app/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 dependencies {
     implementation(projects.feature.home)
     implementation(projects.feature.chat)
-    implementation(projects.feature.add)
+    implementation(projects.feature.newproject)
     implementation(projects.feature.experience)
     implementation(projects.feature.report)
     implementation(projects.core.designsystem)
