+++ app/src/main/java/com/useai/logit/navigation/TopLevelNavItem.kt
@@ -5,10 +5,10 @@ import androidx.annotation.StringRes
 import com.slack.circuit.runtime.screen.Screen
 import com.useai.core.designsystem.R
 import com.useai.core.designsystem.icon.LogitIcons
-import com.useai.feature.add.AddScreen
 import com.useai.feature.chat.ChatScreen
 import com.useai.feature.experience.ExperienceScreen
 import com.useai.feature.home.HomeScreen
+import com.useai.feature.newproject.NewProjectScreen
 import com.useai.feature.report.ReportScreen
 
 data class TopLevelNavItem(
@@ -22,7 +22,7 @@ data class TopLevelNavItem(
             return when (screen) {
                 is HomeScreen -> HOME
                 is ChatScreen -> COVER_LETTER
-                is AddScreen -> NEW_PROJECT
+                is NewProjectScreen -> NEW_PROJECT
                 is ExperienceScreen -> EXPERIENCE
                 is ReportScreen -> REPORT
                 else -> null
