+++ feature/newproject/src/main/AndroidManifest.xml
