+++ feature/newproject/src/test/java/com/useai/feature/newproject/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.useai.feature.add
+package com.useai.feature.newproject
 
 import org.junit.Test
