+++ settings.gradle.kts
@@ -42,6 +42,6 @@ include(":core:ui")
 include(":feature:chat")
 include(":core:navigation")
 include(":feature:home")
-include(":feature:add")
+include(":feature:newproject")
 include(":feature:experience")
 include(":feature:report")
